# api_server/contracts/openapi.v1.json
{
  "components": {
    "responses": {
      "ApiError": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelope"
            }
          }
        },
        "description": "Typed, content-free API error"
      }
    },
    "schemas": {
      "AccessExchangeRequest": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 64,
            "minLength": 8,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "AccessExchangeRequest",
        "type": "object"
      },
      "AccessSession": {
        "additionalProperties": false,
        "properties": {
          "authenticated": {
            "const": true,
            "default": true,
            "title": "Authenticated",
            "type": "boolean"
          },
          "csrf_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Csrf Token"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          }
        },
        "required": [
          "expires_at"
        ],
        "title": "AccessSession",
        "type": "object"
      },
      "ApiErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "minLength": 1,
            "title": "Request Id",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "request_id"
        ],
        "title": "ApiErrorEnvelope",
        "type": "object"
      },
...
      "CreateJobRequest": {
        "additionalProperties": false,
        "properties": {
          "priority": {
            "default": 0,
            "maximum": 10.0,
            "minimum": -10.0,
            "title": "Priority",
            "type": "integer"
          }
        },
        "title": "CreateJobRequest",
        "type": "object"
      },
...
      "JobResponse": {
        "properties": {
          "attempt": {
            "title": "Attempt",
            "type": "integer"
          },
          "cancellation_requested": {
            "title": "Cancellation Requested",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "duplicate": {
            "default": false,
            "title": "Duplicate",
            "type": "boolean"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_retryable": {
            "title": "Error Retryable",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "max_attempts": {
            "title": "Max Attempts",
            "type": "integer"
          },
          "page_id": {
            "format": "uuid",
            "title": "Page Id",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "processed_count": {
            "title": "Processed Count",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          },
          "state": {
            "title": "State",
            "type": "string"
          },
          "total_count": {
            "title": "Total Count",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "document_id",
          "page_id",
          "state",
          "stage",
          "priority",
          "processed_count",
          "total_count",
          "attempt",
          "max_attempts",
          "cancellation_requested",
          "error_code",
          "error_retryable",
          "revision",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "ManualFallbackRequest": {
        "additionalProperties": false,
        "properties": {
          "raw_text": {
            "maxLength": 10000,
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text"
        ],
        "title": "ManualFallbackRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CookieSession": {
        "description": "HttpOnly same-origin session cookie",
        "in": "cookie",
        "name": "htr_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Versioned API v1 contract. Operations marked planned are contracts only and are not exposed by the runtime until implemented.",
    "title": "Tajik HTR Studio API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access/csrf": {
      "post": {
        "operationId": "livePostApiV1AccessCsrfV1",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Csrf",
        "tags": [
          "access"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/access/exchange-code": {
      "post": {
        "operationId": "livePostApiV1AccessExchangeCodeV1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessExchangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "summary": "Exchange",
        "tags": [
          "access"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/access/logout": {
      "post": {
        "operationId": "livePostApiV1AccessLogoutV1",
        "parameters": [
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "access"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/access/session": {
      "get": {
        "operationId": "liveGetApiV1AccessSessionV1",
        "parameters": [
          {
            "in": "cookie",
            "name": "htr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Htr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessSession"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Session",
        "tags": [
          "access"
        ],
        "x-implementation-status": "live"
      }
    },
...
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "liveGetApiV1JobsJobIdV1",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "jobs"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "livePostApiV1JobsJobIdCancelV1",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/jobs/{job_id}/retry": {
      "post": {
        "operationId": "livePostApiV1JobsJobIdRetryV1",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Retry Job",
        "tags": [
          "jobs"
        ],
        "x-implementation-status": "live"
      }
    },
    "/api/v1/jobs/{job_id}/run-regions/{run_region_id}/manual-fallback": {
      "post": {
        "operationId": "livePostApiV1JobsJobIdRunRegionsRunRegionIdManualFallbackV1",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_region_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManualFallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Record Manual Fallback",
        "tags": [
          "jobs"
        ],
        "x-implementation-status": "live"
      }
    },
...
    "/api/v1/pages/{page_id}/recognition-jobs": {
      "post": {
        "operationId": "livePostApiV1PagesPageIdRecognitionJobsV1",
        "parameters": [
          {
            "in": "path",
            "name": "page_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-csrf-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "$ref": "#/components/responses/ApiError"
          },
          "default": {
            "$ref": "#/components/responses/ApiError"
          }
        },
        "security": [
          {
            "CookieSession": []
          }
        ],
        "summary": "Create Job",
        "tags": [
          "jobs"
        ],
        "x-implementation-status": "live"
      }
    },
...
  },
  "servers": [
    {
      "description": "Same-origin deployment",
      "url": "/"
    }
  ],
...
}